Add the basic underpinnings of simply running a SQL script against a Vertica database.
We allow a script to be passed in, which is read, and then using the
specified credentials, connect to Vertica and execute the script without
modification.

We support the notion of swallowing exceptions so that other tasks
requiring this one can avoid having their execution halted if an
optional script run happens to fail.